P_project_with_python/requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
akshare==1.16.91
alpaca-py==0.38.0
alpha-vantage==2.3.1
altair==5.4.1
altgraph==0.17.4
aniso8601==10.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
api==0.0.7
appdirs==1.4.4
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
auto-py-to-exe==2.46.0
autoviz==0.1.905
azure-core==1.30.2
azureml==0.2.7
Babel==2.15.0
backtrader==1.9.78.123
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
bokeh==3.4.2
bottle==0.13.2
bottle-websocket==0.2.9
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
ccxt==4.4.61
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chess==1.11.1
Chessnut==0.4.1
clarabel==0.9.0
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==1.6.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
comtypes==1.4.7
confection==0.1.5
contourpy==1.2.1
cryptography==44.0.1
cvxopt==1.3.2
cvxpy==1.5.3
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
d2l==1.0.0a1
dacite==1.9.2
data==0.4
data-process==0.3.9
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dm-tree==0.1.8
docstring_parser==0.16
duckdb==1.2.0
ecos==2.0.14
Eel==0.18.1
emoji==2.12.1
et_xmlfile==2.0.0
etils==1.9.2
examples==1.0.2
exchange_calendars==4.9
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.11
fastjsonschema==2.19.1
feature-engine==1.8.0
filelock==3.15.1
Flask==3.0.3
flask-cors==5.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.52.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
funcsigs==1.0.2
functional==0.4
future==1.0.0
gast==0.5.4
get-mnist==1.0.1
gevent==24.11.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
gluonts==0.16.1
google==3.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.64.0
gupy==0.0.1
gym==0.17.3
gym-anytrading==2.0.0
gym-chess==0.1.1
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
holidays==0.67
holoviews==1.19.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.0
hvplot==0.10.0
ibapi==9.81.1.post1
idna==3.7
ijson==3.3.0
imageai==3.0.3
ImageHash==4.3.1
immutabledict==4.2.0
importlib_resources==6.4.3
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
iris==1.0.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonify==0.5
jsonpath==0.82.2
jsonpath-ng==1.6.1
jsonpickle==4.0.2
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaggle-environments==1.16.10
kagglehub==0.2.9
keras==3.9.0
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langcodes==3.4.0
language_data==1.2.0
liac-arff==2.5.0
lib-pybroker==1.2.6
libclang==18.1.1
lightning==2.4.0
lightning-utilities==0.14.3
linkify-it-py==2.0.3
llvmlite==0.44.0
lumibot==3.9.17
lumiwealth-tradier==0.1.16
lxml==4.9.4
machine_learning==0.0.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mediapipe==0.10.14
mini-racer==0.12.4
minio==7.2.9
missingno==0.5.2
mistune==3.0.2
ml-dtypes==0.3.2
mnist==0.2.2
more-itertools==10.6.0
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multimethod==1.11.2
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.10
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.2.4
nn==0.1.1
nntplib==0.1.3
nose==1.3.7
notebook==7.2.0
notebook_shim==0.2.4
nsepy==0.8
numba==0.61.0
numpy==1.26.4
numpy4==0.7
numpy6==2.0
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.44.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openml==0.14.2
openpyxl==3.1.5
opt-einsum==3.3.0
optim==0.1.0
optree==0.11.0
osqp==0.6.7.post1
overrides==7.7.0
package_name==0.1
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
pandas-dq==1.29
pandas-profiling==3.2.0
pandas_market_calendars==4.6.1
pandas_ta==0.3.14b0
pandocfilters==1.5.1
panel==1.4.4
param==2.1.1
parso==0.8.4
patsy==0.5.6
peewee==3.17.5
pefile==2023.2.7
pettingzoo==1.24.0
phik==0.12.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
ply==3.11
polygon-api-client==1.14.4
preprocessing==0.1.13
preshed==3.0.9
progressbar2==4.5.0
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyamg==5.2.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycares==4.5.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
pydot==3.0.2
pydotplus==2.0.2
pygame==2.5.2
pyglet==1.5.0
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyluach==2.2.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pypiwin32==223
pyportfolioopt==1.5.5
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytest==8.3.4
python-chess==0.31.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-settings==0.2.2
python-utils==3.9.1
pytorch-lightning==2.4.0
pyttsx3==2.98
pytz==2024.1
pyviz_comms==3.0.2
PyWavelets==1.6.0
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
qdldl==0.1.7.post4
qtconsole==5.5.2
QtPy==2.4.1
Quandl==3.7.0
QuantStats==0.0.62
quantstats-lumi==0.3.3
random11==0.0.1
random2==1.0.2
randomwalk==0.0.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-futures==1.0.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.13.1
scs==3.2.7
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.0.0
shellingham==1.5.4
Shimmy==1.3.0
simple_parsing==0.1.5
six==1.16.0
six21==0.1.3
sklearn01==8
sklearn3==0.0.1
sklearning==0.0.1
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx-rtd-theme==0.2.4
SQLAlchemy==2.0.38
srsly==2.4.8
sseclient-py==1.8.0
stable_baselines3==2.5.0
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.2
stockfish==3.28.0
streamlit==1.39.0
sympy==1.13.1
ta==0.11.0
tabula==1.0.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.6
tensorflow-intel==2.16.1
tensorflow-metadata==1.15.0
tensorflow_keras==0.1
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
thetadata==0.9.11
thinc==8.2.5
threadpoolctl==3.5.0
timeseries==0.5.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.4.0
torchmetrics==1.7.1
torchvision==0.19.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.47.1
tweepy==4.14.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.3
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
utils==1.0.2
uvicorn==0.34.0
visions==0.7.4
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.3
wget==3.2
wheel==0.43.0
widgetsnbextension==4.0.11
wordcloud==1.9.3
wrapt==1.16.0
xgboost==1.6.2
xlrd==2.0.1
xmltodict==0.13.0
xxhash==3.4.1
xyzservices==2024.6.0
yahooquery==2.3.7
yarl==1.9.4
yfinance==0.2.54
zipp==3.20.0
zope.event==5.0
zope.interface==7.2


